chore: rename pages file into one dir
