+++ src/components/pages/HomePage.tsx
@@ -1,12 +1,10 @@
-import { Inter } from "next/font/google"
-import React, { useState } from "react"
+import React from "react"
 import Navbar from "@/components/layouts/Nav"
 import ProductSlider from "@/components/containers/ProductCardSlider"
 import ServiceCardSlider from "@/components/containers/ServiceCardSlider"
 import { products, services, banners } from "@/pages/api/data"
 import BannerCardSlider from "@/components/containers/BannerCardSlider"
 import FooterSection from "@/components/containers/FooterSection"
-import ProductAccordion from "@/components/layouts/ProductAccordion"
 import FormSection from "@/components/containers/FormSection"
 import DesktopFormSection from "@/components/containers/DesktopFormSection"
 import ProductAccordionSection from "@/components/containers/ProductAccordionSection"
