+++ src/components/pages/NewsPage.tsx
